+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Types of changes
 
 ### Added
 
-* Add `l40s_gpu_task` to request L40s GPU instances
+* Add `g6e_xlarge_task`, `g6e_2xlarge_task`, `g6e_4xlarge_task`, `g6e_8xlarge_task`, `g6e_12xlarge_task`, `g6e_16xlarge_task`, and `g6e_24xlarge_task` to request G6e GPU instances.
 
 ## 2.54.10 - 2024-12-06